+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [1.0.1] 2022-07-28
+## [1.1.0] TBA
+
+* fixed safari support for blob file storage by using ArrayBuffer instead
+* added warning when image file url support isn't available because of missing service worker
+
+
+## [1.0.2] 2022-07-28
 
 * Initial release